Separate network protocol from the base data
